rename: domain.Retry => racinggame.RetryInput
